Script for creating fighter json files for autocompete
Since users aren't adding any data we can just store the fighters in memory for autocomplete.
The script makes a list of ufc fighters to be consumed prefetched by the front end. And a json
file of the form 'name':fighterobj to be used with the trie autocomplete in the backend